シンプルなリクエスト/レスポンスができるようになった

+++ modules/domain/src/main/scala/com/louvre2489/fp/domain/value/ValueAdjustmentFactor.scala
@@ -3,6 +3,7 @@ package com.louvre2489.fp.domain.value
 /**
   * 調整係数
   * システムの複雑度などを加味して設定する
+  *
   * @param value 調整係数
   */
 case class ValueAdjustmentFactor(value: BigDecimal) extends AnyVal with Value[BigDecimal]
@@ -11,6 +12,6 @@ case class ValueAdjustmentFactor(value: BigDecimal) extends AnyVal with Value[Bi
   * GSCの評価点
   * @param value 評価点
   */
-case class DI(value: Int) extends AnyVal with Value[Int] {
-  def +(di: DI): DI = DI(this.value + di.value)
+case class DI(value: Short) extends AnyVal with Value[Short] {
+  def +(di: DI): DI = DI((this.value + di.value).toShort)
 }
